+++ views/index.ejs
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Note Taking App - Homepage</title>
+</head>
+<body>
+    <p>
+        Try it out 
+        <a href="/login">Login</a>
+    </p>
+
+    
+</body>
+</html>
