+++ views/login.ejs
@@ -8,6 +8,11 @@
   <div class="container">
     <h1>Login</h1>
 <!-- login.ejs -->
+<% if (message) { %>
+  <div class="alert alert-danger">
+    <%= message %>
+  </div>
+<% } %>
 
 <form action="/login" method="post">
   <label for="email">Email:</label>
@@ -27,35 +32,6 @@
     <p id="error-message" class="error" style="display: none;"></p>
   </div>
 
-  <script>
-    document.getElementById('login-form').addEventListener('submit', async function(event) {
-      event.preventDefault();
-    
-      const email = document.getElementById('email').value;
-      const password = document.getElementById('password').value;
-    
-      try {
-        const response = await fetch('/login', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ email, password })
-        });
-    
-        if (response.ok) {
-          const data = await response.json();
-          document.getElementById('error-message').style.display = 'none';
-          document.getElementById('error-message').innerHTML = '';
-          alert('Login successful! You will be redirected to the dashboard.'); // Add a success message
-          window.location.href = '/dashboard'; // Redirect to the dashboard
-        } else {
-          const errorMessage = await response.text();
-          document.getElementById('error-message').style.display = 'block';
-          document.getElementById('error-message').innerHTML = errorMessage;
-        }
-      } catch (error) {
-        console.error(error);
-      }
-    });
-  </script>
+  <script src="script.js"></script>
   </body>
   </html>
